+++ swap-db/src/alice.rs
@@ -58,22 +58,22 @@ pub enum Alice {
     },
     WaitingForCancelTimelockExpiration {
         monero_wallet_restore_blockheight: BlockHeight,
-        transfer_proof: TransferProof,
+        transfer_proof: Option<TransferProof>,
         state3: alice::State3,
     },
     CancelTimelockExpired {
         monero_wallet_restore_blockheight: BlockHeight,
-        transfer_proof: TransferProof,
+        transfer_proof: Option<TransferProof>,
         state3: alice::State3,
     },
     BtcCancelled {
         monero_wallet_restore_blockheight: BlockHeight,
-        transfer_proof: TransferProof,
+        transfer_proof: Option<TransferProof>,
         state3: alice::State3,
     },
     BtcPunishable {
         monero_wallet_restore_blockheight: BlockHeight,
-        transfer_proof: TransferProof,
+        transfer_proof: Option<TransferProof>,
         state3: alice::State3,
     },
     BtcEarlyRefundable {
@@ -81,21 +81,21 @@ pub enum Alice {
     },
     BtcRefunded {
         monero_wallet_restore_blockheight: BlockHeight,
-        transfer_proof: TransferProof,
+        transfer_proof: Option<TransferProof>,
         state3: alice::State3,
         #[serde(with = "swap_serde::monero::private_key")]
         spend_key: monero::PrivateKey,
     },
     BtcPartiallyRefunded {
         monero_wallet_restore_blockheight: BlockHeight,
-        transfer_proof: TransferProof,
+        transfer_proof: Option<TransferProof>,
         state3: alice::State3,
         #[serde(with = "swap_serde::monero::private_key")]
         spend_key: monero::PrivateKey,
     },
     XmrRefundable {
         monero_wallet_restore_blockheight: BlockHeight,
-        transfer_proof: TransferProof,
+        transfer_proof: Option<TransferProof>,
         state3: alice::State3,
         #[serde(with = "swap_serde::monero::private_key")]
         spend_key: monero::PrivateKey,
@@ -135,7 +135,7 @@ pub enum AliceEndState {
     },
     BtcPunished {
         state3: alice::State3,
-        transfer_proof: TransferProof,
+        transfer_proof: Option<TransferProof>,
     },
     BtcWithheld {
         state3: alice::State3,