+++ swap/src/asb/recovery/cancel.rs
@@ -27,8 +27,10 @@ pub async fn cancel(
         | AliceState::XmrLockTransferProofSent { monero_wallet_restore_blockheight, transfer_proof, state3 }
 
         // in cancel mode we do not care about the fact that we could redeem, but always wait for cancellation (leading either refund or punish)
-        | AliceState::EncSigLearned { monero_wallet_restore_blockheight, transfer_proof, state3, .. }
-        | AliceState::WaitingForCancelTimelockExpiration { monero_wallet_restore_blockheight, transfer_proof, state3}
+        | AliceState::EncSigLearned { monero_wallet_restore_blockheight, transfer_proof, state3, .. } => {
+            (monero_wallet_restore_blockheight, Some(transfer_proof), state3)
+        }
+        AliceState::WaitingForCancelTimelockExpiration { monero_wallet_restore_blockheight, transfer_proof, state3}
         | AliceState::CancelTimelockExpired { monero_wallet_restore_blockheight, transfer_proof, state3}
         | AliceState::BtcCancelled { monero_wallet_restore_blockheight, transfer_proof, state3 }
         | AliceState::BtcRefunded { monero_wallet_restore_blockheight, transfer_proof,  state3 ,.. }