+++ swap/src/asb/recovery/refund.rs
@@ -47,8 +47,10 @@ pub async fn refund(
         | AliceState::XmrLockTransactionSent { transfer_proof, state3, .. }
         | AliceState::XmrLocked { transfer_proof, state3, .. }
         | AliceState::XmrLockTransferProofSent { transfer_proof, state3, .. }
-        | AliceState::EncSigLearned { transfer_proof, state3, .. }
-        | AliceState::WaitingForCancelTimelockExpiration { transfer_proof, state3, .. }
+        | AliceState::EncSigLearned { transfer_proof, state3, .. } => {
+            (Some(transfer_proof), state3)
+        }
+        AliceState::WaitingForCancelTimelockExpiration { transfer_proof, state3, .. }
         | AliceState::CancelTimelockExpired { transfer_proof, state3, .. }
 
         // Refund possible due to cancel transaction already being published
@@ -86,6 +88,10 @@ pub async fn refund(
         let bob_peer_id = db.get_peer_id(swap_id).await?;
         bail!(Error::RefundTransactionNotPublishedYet(bob_peer_id),);
     };
+
+    let transfer_proof = transfer_proof.context(
+        "We have the refund key, but recovery of unknown funds is not yet implemented. Funds are safe.",
+    )?;
 
     retry(
         "Refund Monero",